Adding replace_where incremental strategy (#293)
This PR adds a new incremental strategy replace_where. The strategy resolves to an INSERT INTO ... REPLACE WHERE statement. This completes the feature set explained here: https://docs.databricks.com/delta/selective-overwrite.html#replace-where&language-python

A lot of the code change is to bring macros from dbt-spark over to dbt-databricks. The only real code change was in validating incremental strategies and adding the replace_where strategy.

**Why do we need it?**
It enables use cases where part of the data is always replaced and where MERGE is not possible, such as when there is no primary key. E.g.: events table where we want to always replace the last 3 days.

**Difference from insert_overwrite**
Insert overwrite only works with dynamic partition pruning spark setting, which is not available in sql warehouses or any Unity Catalog-enabled cluster. It also only works with whole partitions, making it difficult to set up and assure that the correct data is dropped.



Signed-off-by: Andre Furlan <andre.furlan@databricks.com>