Updated tests that were modifying the dbt profile
There is a bug in the dbt testing infrastructure: "[CT-3478] [Bug] Editing profiles as part of tests is not thread-safe #9254" https://github.com/dbt-labs/dbt-core/issues/9254
This causes issues for tests that use the profiles_config_update fixture.
Added class level fixture dbt_profile_data to conftest.py.  This creates the profile used by all the tests, in it additional targets are added to be used by the tests that were formerly using profiles_config_update.
Updated the tests to remove the profiles_config_update fixture and to specify the appropriate target when running.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>