Fix: model-level override logic for is_cluster (#1172)
<!-- Please review our pull request review process in CONTRIBUTING.md
before your proceed. -->

Resolves #

<!---
Include the number of the issue addressed by this PR above if
applicable.

  Example:
    resolves #1234

Please review our pull request review process in CONTRIBUTING.md before
your proceed.
-->

### Description

Current logic for `is_cluster_http_path` is backwards. When original
compute is a cluster and the model-level override is a SQL warehouse, it
wrongly returns true. We should use `http_path` as the source of truth
instead of the cluster ID. Cluster ID should only be used as a fallback
if the expected keywords do not exist in the path

### Checklist

- [ ] I have run this code in development and it appears to resolve the
stated issue
- [ ] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.