fix: Diff issues with STs that lead to always triggering config change (#1181)
### Description

In order to not always trigger on_config_change with streaming tables,
we need to track when config changes actually happen; further, as part
of debugging the failing tests, discovered that the Refresh schedule
would always trigger config change due to difference between None and
default of Utc that Databricks stores with.

### Checklist

- [x] I have run this code in development and it appears to resolve the
stated issue
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.