fix: ST set tags were being applied when empty (#1169)
### Description

I don't know how this was happening, since the code for change detection
seems to not return a TagsConfig when I go into via debug, but we were
calling alter with an empty set tags set, which was raising exceptions.

### Checklist

- [x] I have run this code in development and it appears to resolve the
stated issue
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.