breaking: Removing info schema flag (#1226)
### Description

Per discussion with dbt, this behavior is significantly superseded by
DESCRIBE EXTENDED...AS JSON. The only downside is that this behavior is
not available in older DBRs. As such, we will follow up by updating the
warnings in older versions of dbt-databricks that in order to update to
1.11.0, you will need to either use a SQL Warehouse or a relatively
recent version of DBR.

### Checklist

- [ ] I have run this code in development and it appears to resolve the
stated issue
- [ ] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.