refactor: Improve python model tests success rate (#1233)
### Description

So I think the repeated failure of python model tests in our CI/CD is
because we're trying to launch workflows on compute that already has an
overwhelmed driver. For SQL, I think extra statements get queued by sql
gateway, but I think for python models we eventually just time out
saying the cluster couldn't be reached. Also taking this example to
backport AGENTS.md because it's really annoying when working on 1.10.x
stuff to not have the latest changes.

### Checklist

- [ ] I have run this code in development and it appears to resolve the
stated issue
- [ ] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.