fix: Python models over command execution api timeout issue (#1243)
### Description

Command execution api was timing out automatically at 20 minutes due to
defaults with wait. My intent was to grab the command id as soon as it
was available and poll, so that we would have the capability to cancel
if the user killed dbt. Switching back to that approach.

### Checklist

- [x] I have run this code in development and it appears to resolve the
stated issue
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.