fix: Do not set UniForm properties on managed iceberg (#1256)
<!-- Please review our pull request review process in CONTRIBUTING.md
before your proceed. -->

Resolves #

<!---
Include the number of the issue addressed by this PR above if
applicable.

  Example:
    resolves #1234

Please review our pull request review process in CONTRIBUTING.md before
your proceed.
-->

### Description

There is a bug in `TblPropertiesProcessor` that is assuming that all
tables with iceberg format need UniForm properties to be set. When
managed iceberg behavior flag is enabled, this causes the
materialization to fail

The fix is to only apply these properties when working with a UniForm
table. Unfortunately, we cannot use the `file_format` to distinguish
because we do not actually require the user to set `file_format` to
`parquet` for managed iceberg. We infer it in the adapter based on the
`table_format` and behavior flag. To work around this fact, I am setting
`model.config.extra["_databricks_use_managed_iceberg"]` to pass the
information to the config processor

### Checklist

- [ ] I have run this code in development and it appears to resolve the
stated issue
- [ ] This PR includes tests, or tests are not required/relevant for
this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my
change to the "dbt-databricks next" section.